Split audio playback lazy sources
